chore: move quote req validator to util file
